restructuring to make things more stream lined
