+++ lcb/lcb/job_evaluator.py
@@ -1,0 +1,6 @@
+from lcb.evaluator import Evaluator
+
+
+class JobEvaluator(Evaluator):
+    def evaluate(self, args: dict):
+        raise Exception("not implemented")
