+++ lcb/lcb/test_job_evaluator.py
@@ -1,0 +1,26 @@
+def test_evaluate():
+    assert False
+
+
+def test_process():
+    assert False
+
+
+def test_state():
+    assert False
+
+
+def test_download():
+    assert False
+
+
+def test_info():
+    assert False
+
+
+def test_record():
+    assert False
+
+
+def test_exists():
+    assert False
